makefile: run tests inside docker build
If we run the tests before running the container build, the fmt and vet
commands change the last modified times and forces docker to re-add all
the source directories. This is a minor optimization so that if there
are no actual code changes between docker builds the image layers can be
reused from cache.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>