action: copy noobaa images to quay/ocs-dev
Add a GitHub Action to copy the latest noobaa image to quay.io/ocs-dev
for both noobaa-core and noobaa-operator images. When the latest image
is copied, it is also tagged as "master". The master image can be used
in CI to automatically get the latest NooBaa upstream master image.

Run the action every hour. This will make sure that the latest NooBaa
image is copied relatively quickly no matter when it is released, and it
will be retried the next hour if there is a network outage or other
failure. If there are no updates to the image, the action should
complete in 30 seconds or less.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>