controllers: remove setup for StorageCluster controller
Managing the dependent Subscriptions introduced a paradox where we are
trying to watch for a resource type that doesn't exist yet... and
controller-runtime does not like this. This patch temporarily removes
the setup of the StorageCluster controller to work around the CLBO, for
testing purposes, a more appropriate resolution is being developed.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>