controllers: don't set ODF Subscription ControllerReference
Here again we're preparing for OLM to take over managing the overall
state of the Subscriptions. This does leave a TODO for non-RH
Subscriptions, as they will not be marked as dependencies and thus not
removed by the OLM if the ODF Subscription is removed. This is presently
being skipped for the sake of expediency.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>