bundle: bring back dependencies.yaml with better semver
According to the OLM team, the version fields in dependencies.yaml honor
any valid semver formatting. [1] Here we are using it to say the
odf-operator bundle needs *either* ocs-operator 4.9.0 *or* 4.8.z. This
is critically important for upgrades such that the "next" version of
odf-operator (e.g. 4.9.0) can be installed with the "previous" verison
of its dependencies. At this point the odf-operator can take care of
changing the channels in the existing Subscriptions to trigger the rest
of the operator upgrades.

[1] https://devhints.io/semver

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>