hack: add default GOPROXY in Makefile
This is a roundabout workaround for this issue in OLM:
https://github.com/operator-framework/operator-lifecycle-manager/issues/2322

The issue was technically fixed with the v0.19.1 release with the
following PRs:
https://github.com/operator-framework/operator-lifecycle-manager/pull/2331
https://github.com/operator-framework/operator-lifecycle-manager/pull/2336

...however, trying to update the dependency on OLM in ocs-operator
leads to the following issue:
https://github.com/red-hat-storage/ocs-operator/issues/1426

The addition of a default GOPROXY lets us access a cached version of
this package that no longer exists.

Go is wonderful, folks.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>