controllers: start filtering odf sub instead of using static name
While creating a odf-operator subscription from the UI it is always
created by the name of odf-operator. Now managed services is gonna use
the odf-operator as a dependency which means the name will be OLM
generated name and we can not assume static name anymore in the
managed cluster.

Filtering out is a solution to the problem.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>