controllers: stop adding annotation to the storagecluster
adding an annotation to the storagecluster changes the resource version
of the storagecluster on etcd and ocs-operator fails to update the
status while onboarding the consumer.

There is no easy way to get the id of the storage consumer from the
provider again after the onboarding is successful from the provider end.

moreover annotation was used to skip the reconcile which can be done via
predicates also. using create predicate will act only once unless pod is
restarted. It will also act on already existing storagecluster.

minimize the update calls wherever possible.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>