test: Check for all the required CSVs after odf operator is installed
We check for all the CSVs, if they have succeeded or not after the
odf operator install is done in the Setup function in the BeforeSuite.
The CSV names to be checked are passed while making the e2e-test.
Also as the all the CSVs are checked here so the waitforcsv call
while Creating olm resources has been removed.
Signed-off-by: Malay Kumar Parida <mparida@redhat.com>