bundle: Add csi-addons to the dependencies.yaml
csi-addons resolution gets fail when user uses the custom catalog
source as it is not specified in the dependencies.yaml and the sub is
not created via OLM. The sub is created by the odf directly which does
not have a knowledge about the custom catalog source and it is still
pointing to the redhat-operators in the openshift-marketplace namespace.

Managing there own catalog source is seen very offently now a days as it
is MS and disconnected ENV usecase.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>