action: do not run catalog build using local registry
As opm render is running inside a container while making a container
image it does not have a access to the local registry and fails to
render odf-operator-bundle image.

To render odf-operator-bundle we will have to push it to the internet
where it is publically accesable or we can use the opm binary directly
on the machine instead of using it inside container.

It will push it to ocs-dev for now.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>