makefile: remove OPERATOR_NAMESPACE env var from the configmap
Adding the OPERATOR_NAMESPACE constant value to the configmap causes few
issues if we deploy the operator in any other namespace. Because this
value is set at the build time, it cannot be changed while the operator
is being deployed. Instead, we'll include this environment variable in
the deployment and set its value using the downward API.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>