bundle: Modify the name and reference of the sample storagesystem
In certain instances customers unintentionally create two storagesystems
to the same storagecluster. This becomes problematic when attempting to
delete one storagesystem, as it triggers the deletion of the associated
storagecluster. By changing the name and reference, we aim to prevent
this issue from occurring.

It will make sure that even if they create the storagesystem from the
OCP console it wont refer to the same storagecluster.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>