godeps: fix an error 'unrecognized import path "go.opencensus.io"'
github.com/census-instrumentation/opencensus-go repository was archived
on the Jul 31, 2023 and url go.opencensus.io holding these libraries is
not available anymore. Though it is still available on the proxy server.

But we are using GOPROXY="direct,https://proxy.golang.org" which tries
to fetch it from the go.opencensus.io and throws an error. To fix this
we are now starting pulling it directly from the github archived repo.

once we start using GOPROXY="https://proxy.golang.org,direct" we can
revert this change as well.

while reverting 6cb13e914938e7c8d2832fde0b930cebcb84a98e we also need
to revert this.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>