controllers: add custom label to odf CSV to skip webhook call
Relying on the OLM label was unreliable as it gets added after CSV
creation, causing the webhook to trigger for odf-operator CSV.

This change adds a custom label to ensure the webhook skips the
odf-operator CSV during creation.

It prevents issues where uninstalling and reinstalling ODF blocks
odf-operator CSV creation due to webhook CR presence in the cluster,
while the odf-operator is absent to serve the webhook requests.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>