controllers: fix replica update logic for CSVs during upgrades
Previously, the controller would list all CSVs with a package label
and update their replicas. This caused issues during upgrades because:

- It would scale up the ocs-client-operator 4.18 CSV.
- The ocs-client-operator 4.18 CSV pods would get stuck in init state.
- This blocked the upgrade of ocs-client-operator the 4.18 CSV remained
  in Installing state while 4.19 CSV was stuck in Pending.

The new implementation specifically targets the correct CSV to avoid
interfering with Previous CSV during upgrades.

This occurred because odf-operator 4.19 gets upgraded before other
operators and start modifying the 4.18 CSVs.

The problem is happening only with the ocs-client-operator as it was
scaled down by odf-operator in 4.18 and previos releases.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>