makefile: use the per-release CM to prevent stale reads during upgrade
Before this change, the odf-operator read from the ConfigMap named
odf-operator-manager-config before performing any action. During the
upgrades, the configmap gets updated before the operator is upgraded.
As soon as the configmap gets upgraded operator will act on it and
start reconciling the resources based on the updated configmap.
The updated configmap contains the new values that the operator
should not reconcile.

To address this, this PR introduces a dedicated ConfigMap for each
release. The odf-operator now references the ConfigMap specific to
its version. During upgrades, both the old and new ConfigMaps coexist
in the cluster, ensuring that the old operator instance reads only the
configuration intended for its release.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>