controllers: Skip cleanup controller on fresh installs
Previously, creating the cleanup controller on fresh installs where the
StorageSystem CRD is not present caused below issue:

Error: if kind is a CRD, it should be installed before calling Start

To address this, we now check for the presence of the StorageSystem CRD
before creating the cleanup controller. If the CRD is not found, we skip
creating the controller, avoiding errors during fresh installations.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>