controllers: Improve OpenShift version detection logic
The clusterVersion CR has openshift version in 2 places,
- Status.Desired.Version
- Status.History[].Version
During a version upgrade, the desired version is updated first although
the upgrade is not yet completed, and the upgrade may not complete ever
in some cases. So the true version is the latest completed version in
the history.
Also in practice there is only one cluster version CR with name
"version", hence no need to do a list.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>