controllers: Make operator not upgradeable if OCP upgrade is incomplete
If the OCP upgrade is incomplete, the ODF upgrade is not safe to
proceed. Hence, mark the operator as not upgradeable to ensure cluster
stability.

To resolve this user would have to either wait for the OCP upgrade to
complete or override the operator upgradeable condition via the
operatorCondition CR.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>